Add notebooks and script
One notebook is used for gathering historic estimates based
on the state of Commons as that time. The other is for the
weekly updates of the data, which is also published to HTML
and copied to be publicly available. Lastly, the script that
runs as a cron job once a week to do the updating.